requirements.txt
streamlit
PyPDF2
pdf2image
pdf2docx
PyMuPDF
pdfminer.six
pillow